refactor(ui): extract tool-info helpers to dedicated module
Slice 11b.1 step 4/14 of the message-part.tsx split. Moves the i18n-
aware tool-info factory (`getToolInfo`) and the per-tool state
narrowing helpers (`toolStateMetadata`, `toolStateError`) out of the
monolith into `message-part-tool-info.ts`, alongside the agent tone
palette + `taskAgent` resolver that the agent/task tool renderer uses.

These helpers are all "pure (per render scope)" — `getToolInfo`
internally calls `useI18n()`, so it must remain inside a Solid render
scope; it has not become module-level callable.

The public `buildToolInfo` / `ToolInfo` re-export now flows from
`./tool-info` directly (its original source) rather than transiting
through message-part.tsx — equivalent behavior, smaller import graph
inside the new tool-info module.

message-part.tsx 2263 → 2103 lines (-160). Typecheck GREEN, 478 pass /
3 baseline fail unchanged.
